chore: bump kommander chart to 0.4.32
The 1.0.x release of this addon was using the 0.4 version of the chart.
Konvoy shipped with 1.0.2 which used chart 0.4.31. Since upstream helm
charts are removing their stable repository from google, we need to
point the subcharts of that chart to our own fork of the upstream repo.
This requires using the 0.4.31 chart and only changing the repo url in
requirements.yaml.